light-base: move the lifecycle tasks into lifecycle_service
